refactor: 문제 도메인 서브 패키지로 이동
